Tagging version 2.2
git-svn-id: https://plugins.svn.wordpress.org/duplicate-post/tags/2.2@484998 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ duplicate-post-admin.php
@@ -43,18 +43,9 @@ function duplicate_post_plugin_upgrade() {
 			$role->add_cap( 'copy_posts' );
 		}
 			
-		add_option(
-		'duplicate_post_copyexcerpt',
-			'1',
-			'Copy the excerpt from the original post/page' );
-		add_option(
-		'duplicate_post_copystatus',
-			'0',
-			'Copy the status (draft, published, pending) from the original post/page' );
-		add_option(
-		'duplicate_post_taxonomies_blacklist',
-		array(),
-			'List of the taxonomies that mustn\'t be copied' );
+		add_option('duplicate_post_copyexcerpt','1');
+		add_option('duplicate_post_copystatus','0');
+		add_option('duplicate_post_taxonomies_blacklist',array());
 	} else if ( $installed_version==duplicate_post_get_current_version() ) { //re-install
 		// do nothing
 	} else { //upgrade form previous version
@@ -70,37 +61,30 @@ function duplicate_post_plugin_upgrade() {
 		
 		// Get old duplicate_post_copy_user_level option
 		$min_user_level = get_option('duplicate_post_copy_user_level');
-		
-		// Get default roles
-		$default_roles = array(
+
+		if (!empty($min_user_level)){
+			// Get default roles
+			$default_roles = array(
 			1 => 'contributor',
 			2 => 'author',
 			3 => 'editor',
 			8 => 'administrator',
-		);
-			
-		// Cycle all roles and assign capability if its level >= duplicate_post_copy_user_level
-		foreach ($default_roles as $level => $name){
-			$role = get_role($name);
-			if ($role && $min_user_level <= $level)
-			$role->add_cap( 'copy_posts' );
+			);
+				
+			// Cycle all roles and assign capability if its level >= duplicate_post_copy_user_level
+			foreach ($default_roles as $level => $name){
+				$role = get_role($name);
+				if ($role && $min_user_level <= $level)
+				$role->add_cap( 'copy_posts' );
+			}
+
+			// delete old option
+			delete_option('duplicate_post_copy_user_level');
 		}
 		
-		// delete old option
-		delete_option('duplicate_post_copy_user_level');
-		
-		add_option(
-		'duplicate_post_copyexcerpt',
-			'1',
-			'Copy the excerpt from the original post/page' );
-		add_option(
-		'duplicate_post_copystatus',
-			'0',
-			'Copy the status (draft, published, pending) from the original post/page' );
-		add_option(
-		'duplicate_post_taxonomies_blacklist',
-		array(),
-			'List of the taxonomies that mustn\'t be copied' );
+		add_option('duplicate_post_copyexcerpt','1');
+		add_option('duplicate_post_copystatus','0');
+		add_option('duplicate_post_taxonomies_blacklist',array());
 	}
 	// Update version number
 	update_option( 'duplicate_post_version', duplicate_post_get_current_version() );
@@ -245,7 +229,9 @@ function duplicate_post_copy_post_meta_info($new_id, $post) {
 	foreach ($meta_keys as $meta_key) {
 		$meta_values = get_post_custom_values($meta_key, $post->ID);
 		foreach ($meta_values as $meta_value) {
-			add_post_meta($new_id, $meta_key, $meta_value);
+			$meta_obj = unserialize($meta_value);
+			if(!$meta_obj) add_post_meta($new_id, $meta_key, $meta_value);
+			else add_post_meta($new_id, $meta_key, $meta_obj);
 		}
 	}
 }
@@ -286,7 +272,6 @@ function duplicate_post_create_duplicate($post, $status = '') {
 
 	$new_post_id = wp_insert_post($new_post);
 
-	add_post_meta($new_post_id, '_dp_original', $post->ID);
 
 	// If you have written a plugin which uses non-WP database tables to save
 	// information about a post you can hook this action to dupe that data.
@@ -294,6 +279,9 @@ function duplicate_post_create_duplicate($post, $status = '') {
 	do_action( 'dp_duplicate_page', $new_post_id, $post );
 	else
 	do_action( 'dp_duplicate_post', $new_post_id, $post );
+	
+	delete_post_meta($new_post_id, '_dp_original');
+	add_post_meta($new_post_id, '_dp_original', $post->ID);
 
 	// If the copy gets immediately published, we have to set a proper slug.
 	if ($new_post_status == 'publish' || $new_post_status == 'future'){
